Use version from module, not txt file
this is cleaner since we don't have to read .txt files and don't
modify the source before building a package.

This makes it possible to show the actual version in Nitrate's webUI
footer by exporting the version as a setting.

The setting can be overriden if desired, for example with
commands.getstatusoutput('rpm -q nitrate')[1]

Signed-off-by: Mr. Senko <atodorov@mrsenko.com>