Let testcases tests run again
Test case CaseTests is deleted. At the present time, I have no idea why
that case was written in that way, and those tests are not clear enough
to tell what they test. In addition, the assertions don't make sense.
It is not worth to spend too much time on this case to make them run
again. So, just remove this test case simply, and write new test cases
for models, views (only the necessary part of views it's worth to
test), and other part of testcases app.

Related to #113

Signed-off-by: Chenxiong Qi <qcxhome@gmail.com>