Implement the 'Show all bugs URL' option in TestRun report
Using the new issuetracker integrations code we produce a list
of URLs which can load multiple bug IDs into the IT interface.
This is supported for Bugzilla and JIRA but may not be supported
for other issue trackers.

If the issue tracker is not configured for report urls nothing will
be shown even if that particular type of IT supports the
functionality.

This commit takes care to identify unique bug trackers which may
have been used for recording bugs inside the TestRun. We produce
a reporting URL for every issue tracker we can.