Remove integration with Errata System. Fixes #15
An Errata System is something which hasn't been documented since
the early days of this project and I guess is something internal
to Red Hat. This commit removes references to this system and also
removes the tcms.integrations.errata application together with its
messaging layer. If need be messaging can be brought back together
and refactored to be more versatile.

- removed configuration setting ERRATA_URL_PREFIX
- adds migrations/0004_remove_testrun_errata_id.py