Drop dependency on Kobo. Fix #5
- copy XML-RPC related code from Kobo
- drop the dependency

Note: Kobo appears to be an old piece of code which isn't actively
maintained with latest Django versions and appears to include
bunch of random integrations from Red Hat's Release Engineering.
I've taken only the bare minimum code which is relevant to us
and copied it here so we can drop the dependency.