Don't wrap Prompt.render with HttpResponse
Prompt.render calls django.shortcuts.render which returns an
HttpResponse. When HttpResponse.__init__ is executed the content
is assigned to self.content via a setter method. This method
accepts either string content (rendered from templates) or
iterator.

When we wrap Prompt.render with additional HttpResponse the
retrurned content is the same but because now we pass an iterator
to the __init__() method another branch of the code is executed.
This branch calls content.close() which in this case is
HttpResponse.close() which fires `signals.request_finished` which
in turn closes the connection to the database.

During testing this is a problem because the connection is closed
right after the offending view returns, which is in the middle of
a test case method. Because the DB connection is now closed all
subsequent queries that are used to assert the application state
(e.g. User.objects.count()) fail with TransactionManagementError!