Use min-height CSS property for toolbar buttons
these are the 2nd tab buttons inside the main content, e.g.
for test plan actions. Switching to patternfly caused then to be
displayed with height 4px which broke the popup menus because
all of them use absolute coordinates.

All of these issue will be fixed when these components are migrated
to patternfly as well.