Load admin/js/vendor/jquery
this commit loads the jquery version from the admin so that the
changes in e314ffe can possibly work. It also cleans up a bit
the admin templates.

NOTE: the {{ media }} tag will load js/css from admin forms
defined inside Django and some of these are also manually loaded
in admin/base_site.html and tcms_base.html as well. IMO this is
a mess and should be cleaned up during the patternfly migration!