If JIRA isn't fully configured then don't connect. Fixes #100
turns out the underlying jira library tries to connect
automatically when a jira.JIRA object is created. When the
Issue Tracker is not fully configured (i.e. missing base_url,
api_url, api_username, api_password) then the library will default
to localhost and fail to connect to the JIRA server.

This causes the TestRun report view to timeout and fail.

This commit adds tests for all supported issue trackers and also
fixes the JIRA implementation to not connect if not configured.