Remove try/except/raise kerberos.GSSError
so we catch a kerberos.GSSError with except and then raise the
same type of error as the handler!?! I don't see the point in
doing that when we can simply let the exception escape!

Removed the redundant try/except/raise blocks because of that.
Also removed white space around those lines.