Disable API cookie transport on Python 3
In Python3 the single_request() method just calls other
methods to do the sending and acts more like a high-level method
compared to Python 2. This also makes it harder (for now) to
implement cookies without rewriting lots of the XMLTransport
functionality.

Disable cookies for now on Python3 and figure out how to deal with
that later!