Log XML-RPC requests from anonymous users. Fixes #126
we acomplish this by creating (if necessary) a special user object
with username=='Anonymous'.

Note: While it is possible to log the requests with user=None
the admin interface needs to be modified to support
displaying/filtering None values which is even more hassle.
This is the lest intrusive way to fix this problem!