git ignore package-lock.json
This is breaking docs tests in Travis CI but at the same time
on RHEL 7 or CentOS 7 is not an issue because we're using an older
npm there. See:
https://docs.npmjs.com/files/package-lock.json

NOTE: The idea of this file is to pin-down all version in the
dependency tree, which is a good idea by itself. For the moment
we've not seen issues with versions changing underneath us so we're
going to ignore this for now!