[bug-fix] Properly urlencode emails in personal menu
when email contains + sign, e.g. user+qa@example.com this was causing
the search views to not find any Test Plans or Test Runs. The reason
is because the email address needs to be properly urlencoded before
passing it as query string argument to the search views!

Without urlencoding the + sign is then decoded as a space within
the search view!