[tests] RPC logging test now sends requests through LiveServer
this is a more accurate representation of how things work in
real-life! Previously the home-grown code was doing lots of this
behind the scenes but now if we want to test how an RPC method
behaves (as seen from the client) we need to serve it via LiveServer
and call it via a client-side XML-RPC object.

Also removed unnecessary functions for XML-RPC tests which will be
refactored.