Hide admin forms for some models. Fixes #174
Django admin uses generic, automatically generated forms for
the models. This is in conflict with the specialized add/edit
forms and views that are used for things like Test Plans for
example.

There are also models like TestCaseStatus and TestCaseRunStatus
whose records have special meaning inside the application and
should not be modified.

The special Sites object is used to configure the fully qualified
domain name of the running Kiwi instance and nothing more so we
redirect Add/Delete operations to the Edit operation for the
currently configured object.