Correctly find property ID when renaming. Fixes #167
the problem was that for newly added property the 'Rename' action
only searched its immediate parent while the object ID was stored
one level up. This commit updates how we search for the parent.

Also update the template for newly added property to match the
existing HTML for other properties!