Convert request body to string. Fixes #177
request.body returns bytes, e.g. b'abcd'. This is replaces into the
template as-is and later the JavaScript code works with it, trying
to serialize the search form and any previous search criteria when
doing bulk update actions like changing a TC status!

Because the form serialization operations are home-drown the
existing b' is left and subsequently it breaks elsewhere.

In this particular case this leads to page a page size parameter
posted with value 20' which breaks integer conversion.

The UI serializing forms and posting www-urlform-encoded to the
server is a total mess. We should instead post pure JSON and
receive pure JSON.