Remove use of Prompt.render for get, print and export views
For print and export there is a JavaScript function which
alerts the user if no objects were selected. Inside get() the
prompt was shown if the specified from_plan parameter didn't
match an actual TestPlan. We default to setting the test plan
to None as is the case when no test plan is specified.