Remove testplans.views.delete and friends
This view function has never been used. It is not possible to
delete TestPlans via the regular UI and we've disabled access to
TestPlan admin in bf2b4952c9de582cc49b4925bf5fd05060e858ad.
The only way to actually delete a TestPlan object is via the
Django shell! If test plan is not to be used it should be disabled
instead!

With this view gone we also remove:

- PLAN_DELELE_EMAIL_TEMPLATE setting
- mail/delete_plan.txt template
- handle_emails_pre_plan_delete signal handler
- TestPlanEmailSettings.notify_on_plan_delete model field