Move SSL keys under common directory in the container. Fix #231
The problem is that some older versions of docker do not allow
mounting of files between the host and the container, they only
allow mounting directories and volumes. The stock docker versions
on CentOS 7 and RHEL 7 do this. You may see a similar error:

ERROR: for kiwi_web Cannot start service web:
    OCI runtime create failed: container_linux.go:348:
        starting container process caused "process_linux.go:402:
            container init caused "rootfs_linux.go:58: mounting
                "/root/kiwi/contrib/ssl/kiwi-https.crt" to
                rootfs "/var/lib/docker/overlay2 ....

To work around this problem I've moved the ssl certificates into
a common directory, mount that into the container and updated the
file paths where necessary.

WARNING: If you decide to override settings you will have to either
mount the settings file into the container or COPY that file and
rebuild the image!