Expose httpd logs to the host running docker. Fixes #191
this commit makes it easier to access the httpd logs, mostly useful
in case of error, and adds a small section about troubleshooting in
production.

There could still be issues with SELinux and file ownership depending
on how the host setup.

The helper instroduced in this commit only gets applied if you
use docker-compose to start Kiwi TCMS. If you don't logs will still
be inside the container!