Remove TestRun cloning from search results and plan details
If you want to clone an existing TestRun then you should open it
and use the Clone button!

This commit helps us get rid of duplicate clone() view which is
used in some cases, while the new_run_with_caseruns() method is
used in other cases.

Also remove other code which is solely related to the functionality
being removed!