Globally disable delete_selected action in Admin. Fixes #221
this removes the drop down selection widget!

NOTE: we could add a button in template's `object-tools` block
but that will not be able to handle multiple selection. Also the
checkboxes for multiple selection are not rendered when
delete_selected is disabled. For reference see:
https://github.com/sehmaschine/django-grappelli/issues/839

If the user wants to delete an object they will have to open each
object in a separate page and delete it from there!