[bandit] Remove veiw that calls exec & import with untrusted data
Issue: [B102:exec_used] Use of exec detected.

Fix: Completely remove the view which does this. That was meant as
     a function which will dynamically return HTML forms for
     rendering on the front-end by receiving form module.ClassName
     from the sender!

     The offending code was blindly trusting untrusted input and:

     exec('import tcms.%s as form' % request.GET.get('app_form'))
     __import__('tcms.%s' % request.GET.get('app_form'))

     This is actually a big deal because it allows remote code
     execution by sending a very simple POST request!

Because the offending view is actually always called with a single
value I replace it with a view that returns that particular form
as HTML and also update the JavaScript code.