Allow to delete TestPlan. Fixes #273
user is redirected to the admin view for deletion. The admin view
is limited on purpose so that only delete action can be performed.

On the TestPlan view page there is a DELETE PLAN link at the top.
It is rather ugly one and will be restyled when we convert this
template to Patternfly.