Rename tables in hard-coded SQL
after the rename of the DB tables we have to rename these
hard-coded SQL queries because otherwise they fail. These will be
removed soon so a quick fix is acceptable.

NOTE: as these are not thorughly tested they could lead to bugs
but we'll worry about that if/when we get there.