Don't track & display history for EnvGroup model
this doesn't seem to be a very useful feature for the moment and
if necessary it should be implemented via simple_history or
model_utils instead of home grown code.

This is part of simplifying and removing the home-grown logging
classes before we can migrate to something else.

Also remove the author and modified_by fields because we don't
really care who created this record and modification will be
recorded via history.