[ui] Remove Disable/Enable Plan buttons
These buttons are backed by the deprecated ajax.py:update view
which we'd like to remove or at least not use as much as possible.

Given that enabling/disabling can be done by editting the TP and
clicking the Active checkbox I don't see a reason why we should
have these 2 buttons in a prominent place.

Plus as it is enabling and disabling doesn't keep track of history!