Remove TCMSActionModel.log_action()
this is now mostly duplicated by simple_history. What is missing
is the M2M history which is still in progress upstream!

However that may turn out not so useful. The code for showing
M2M history (e.g. cases being added to plans) will be removed in
the next commit. If it turns out we need it it should be implemeted
via simple_history!