Convert TestCaseEmailSettings.cc_list to string field
instead of working with custom objects and saving each address
as a separate record into the DB we convert this field to TextField
and save everything together. The underlying interface for working
with CC addresses already provides a list interface so we keep that.

Also removed now unused models and functions.