Update TestRun complation status manually. Fix #367
TestRun.update_completion_status() is now called in only 2 places
after a user selected action. We've removed the functionality to
update the status automatically and removed all signals related
to that.

This will also fix #367 because emails are sent after TestRun.save()
which now is less often (e.g. when user edits or changes status).