Rename tcms.core.contrib.auth app to tcms_auth
this is required because ./manage.py dumpdata/loaddata expects
the models to be named "app_name.model_name" and aprsing fails when
there are multiple dots in the name.

Because the name has changed a new DB table will be used for the
UserActivationKey table and we remove the table with the old name.