Update has_change_permission() and don't allow others to change passwords
this will hide the list of all users from non-superusers but
still allow them to change their passwords or view profiles.

There are 2 forms which allow changing passwords in Django Admin

1) change the password for a given user if you have permissions
2) change your own password by confirming the current one

This commit changes the UserAdmin interface so that all links lead
to form 2). If the user is unable to change their own password
they can always try to reset it.