Add JavaScript JSON-RPC client
this will be used to communicate with the existing JSON-RPC API
provided by Kiwi TCMS. This is the same interface as the XML-RPC
API only served via different endpoint and protocol.

The intention is to replace all 'ajax' views in favor of the
existing API and not duplicate functionality simply to make
a dynamic front-end.