Remove unused dataTables.numHTML.js
previously this was used to correctly parse ID numbers from
tables and the only place where it was used was on
TestCase page -> Test Plans tab. However we don't show the IDs
in this table anymore so the entire column needs to be removed and
there's no need for this extension.