Submit filter form when clicking on tag value. Fix #426
NOTE: in TestPlan page, Cases tab, when clicking on a "tag"
it will automatically update the filter criteria and submit the
filtering form which will refresh the tab with new values.

This will also update the list of available tags to the
currently chosen one, b/c this list is calculated *after*
test cases are filtered (which depends also on the selected tag)!