Don't use |slugify filter in templates. Fix Sentry KIWI-TCMS-38
the problem is that in code we use uuslug.slugify while in templates we
use the slugify filter which is django.utils.text.slugify and that skips
non-ascii charaters. So when resolving URLs (to test plans) that have the
slug in the address (for human readable URLs) we get an error like:

Reverse for 'test_plan_url' with arguments '(1034, '')' not found.
1 pattern(s) tried: ['plan/(?P<plan_id>\\d+)/(?P<slug>[-\\w\\d]+)$']

https://sentry.io/open-technologies-bulgaria-ltd/kiwi-tcms/issues/676626096/