Add tests for tag permissions for TestPlan API and view
also update the template so it inspects the correct permissions
before showing the 'Add Tag' widget. management.add_tag controls
if the user can add Tag objects to the DB not if they can assign
tags to TestPlan or TestCase objects!

Test with all possible permissions minus the ones we are testing
for to catch scenarios where the application is examining the
wrong permission!