Update Changelog for the interim 6.0.1 release
direct upgrades from 5.3.1 to 6.1 work fine but
direct upgrades from 6.0 to 6.1 fail because of the
order in which migrations are executed.

The problem is that we add new migrations which remove fields
from the DB and later (in the same release) squash them with
previous ones which messed up the Django planner.

6.0.1 has no functional differences, it was just tagged from
a previus commit.