On dashboard show TRs which have cases assigned to current user
Fix #520.

Use case:
We usually create one TestRun for one version of the product
which contains lots of cases assigned to multiple testers.
To simplify the procedure for testers (to avoid searching for specific
TestRun or having to save direct URL to it) it would be beneficial
to show link to Test Run in Test Executions section on Dashboard
to all users which are assigned to at least one test case in that
Test Run.