Hide KiwiUserAdmin.password field from super-user. Refs #610
override the fieldsets even if viewing the User admin form as
super-user so that we hide the password field. We do this b/c
this field has a link to change the password, which doesn't
change the password for the selected user but for the super-user
itself. Hide the field so it is not confusing.