By default expire session cookies after 24hrs. Fixes #556
this is mostly used to document the correct setting so that people
who are concerned about it can find it more easily and override it!

Whatever default value we set as the default will not suit everyone
so I go with a value that I like. In practice you should be serving
Kiwi TCMS over https and in controlled environment so this should
be less of an issue.