Quick fix: when viewing TestPlan cases make TC summary link to TC
previously the summary column was a link going to nowhere. It was
intended to be clicked with the mouse and expand TestCase details
the same way clicking the blue arrow does. This wasn't working or
at least hasn't been working for some time! Not sure if it was
always that way or we broke it somehow.

Anyway this entire template will be removed once we migrate the
page to Patternfly and start using the TestCase search component!