Refactor removeCaseRunBug() to JSON-RPC, Refs #18
the UI updates are a bit tricky b/c we still show bugs that have
been reported in other TestRuns or generally attached to the
TestCase. The idea is that these can be known issues and should be
easy to find when testing (i.e. we're seeing this behavior for the
first time but somebody else had reported it already).