Test + fix for failing TestCase rendering in French
when displaying a TestCase with the new template we show icons
based on execution status by using a dict lookup. Trouble is
these statuses can be translated and the dict lookup will fails
if the user has their browser configured for soemthing else than
English.

Note: to properly be able to test with different languages and
discover all places where we may have the same problem we need
this feature from Django:
https://github.com/django/django/pull/10988