Replace |stringformat with |escape so selected option is shown!
There was a problem with |stringformat where it worked when there
were errors in edit mode but not when there were errors when
creating a new object.

I'm not exactly sure why but |escape seems to work and I've used
it elsewhere. I think it works because we convert both sides of
the comparison operator to strings regardless of what their initial
types are!

@asankov CC