[db] populate missing TestCase.text history
in version 6.5 the TestCase model was updated to store the text
into a single field called `text` instead of 4 separate fields.
During that migration historical records were updated to have
the new `text` field but values were not properly assigned.

The "effect" of this is that in TestCaseRun records you were not
able to see the actual text b/c it was None.

This commit ammends 0006_merge_text_field_into_testcase_model for
installations which have not yet migrated to 6.5 or later. We also
provide the data-only migration 0009_populate_missing_text_history
which will inspect the current state of the DB and copy the text
to historical records.