Fix visualisation of X axis for ExecutionTrends
Turns out `C3ChartConfiguration.config.axis.x.tick.format` is not working as expected, as the first argument passed to the function is the order of the entry, not the entry itself.
Instead, we are now mapping the entry from the data array (test run ID) to the wanted format ("TR-{id}") by hand in the code.
Also, use `Array.prototype.map` to do the mapping and return the result array directly, instead of creating an empty array and filling that in `Array.prototype.forEach` loop