Clone TP: always 'Copy All Test Cases' & keep default tester
the original label of this checkbox is misleading. It was meant to
mean that all test cases from the source TP will be included in the
destination TP.

In addition this field was named link_testcases which together with
copy_testcases were controlling if TCs will be just added to the
new TP (aka link via m2m relationship) or cloned with new TC PKs.
The two options are opposite so it doesn't makes sense to keep
only 1 of the two boolean fields to control this behavior.

Always keep the Default tester for test cases