[API] New method Bug.details(). Refs #117
this method together with the underlying IssueTracker.details()
will be the foundation of how Kiwi TCMS fetches more details from
the bug tracking system.

At the moment the intention is for this information to be shown in
the UI but it can also be consumed independently via the API.

By default we're using OpenGraph protocol to collect the data that
will be shown. The details() method can be overriden for each bug
tracker that is supported and you may also override the entire class
with something you have created yourself.