Don't start/end translations with newline character. Refs #1098
Crowdin in-context translation for Esperanto uses automatic msgstr
which don't have these characters. This in turn causes
compilemessages to fail!

Note: this is mostly an issue where we use blocktrans and multi-line
strings. The solution is to make them look uglier :-(. In the final
HTML/email they will be displayed the same.