tests: assert based on translated string
b/c updated in French translation later.

Note: the actual French string is
"L'utilisateur saisi n'existe pas en base" where the
apostrophes get rendered as &#39; in HTML and although the result
is translated the assertion fails.

As a workaround we'll not merge the commit introducing this
string ATM.

CC @RMadjev