Document bug tracker integration support. Fixes #698
as of now this is:

- 1-click bug reporting with fall back of opening URL with
  pre-filled information where possible
- "linking" TestExecution with existing bugs, usually by adding
  a new comment to the bug pointing back to a failed TE
- showing bug details in a tooltip for pages which display bugs

There have been no other ideas for "different types of integration"
so I'm calling #698 closed!

What remains are bug fixes in existing functionality and/or
partially implemented integration/integration which can be improved
for 3rd party bug trackers. These will be solved in their respective
bugs.