Document timezones. Closes #1206, Refs #545
all of these settings can be customized by the administrator
running Kiwi TCMS. The only thing we can't easily do is display
the data in different time zones when hosting an application for
geographically dispersed team. For this reason the previous commit
displays the server time in the navigation bar.