Update MultipleEmailField.clean() to return a string
this should have been like that from the start but b/c we've been
using this field manually in update_case_email_settings() we
haven't noticed the issue.

The DB store is a TextField so .clean() should return a string
instead of a list! With the introduced usage of InlineModelForm
in the previous commit this problem became obvious!