[security] Add JSON-RPC handler which will HTML escape all strings
this prevents XSS attacks via tags, components or anything else
which is loaded on the web page via RPC and then shown as string.

Even if someone saves <script>alert(123);</script> in the DB
the returned result will be HTML escaped and will not be executed
as JavaScript!

Severity note:
this is easy to exploit but people able to do so should have
accounts in your Kiwi TCMS installation and write privileges on
their accounts. If they do this means they can do a lot more damage
much more easily.

Other notes:
- https://public.tenant.kiwitcms.org has been patched
- brought to my attention by @Prome88