Define TENANT_APPS b/c impossible to check for errors otherwise
this is a setting which is relevant only for
kiwitcms-tenants/django-tenants and for Kiwi TCMS represents a
subset of all applications accessible to each tenant. These are
the appilications which hold private data for that tenant.

In the past we've had issues where adding a new app into
INSTALLED_APPS (e.g. tcms.bugs) didn't make it into TENANT_APPS
and resulted in all users being able to share bugs which we didn't
want.

Because TENANT_APPS is always a subset of INSTALLED_APPS it is not
possible to check if we've added something to INSTALLED_APPS but
forgot to add it into TENANT_APPS. That's why I'm moving this
setting here and using it as a temporary variable to construct the
full list of INSTALLED_APPS.