Revert rich HTML bug details for internal bug-tracker
Since we started HTML-escaping all strings returned via JSON-RPC
in order to stop XSS vulnerabilities there's no good & secure way
to return HTML from the BE directly to the browser.

This commit makes the tooltip less verbose which makes it display
better in the browser but is just a workaround!