Unescape &amp; back to & when opening new windows. Fixes #1533
the underlying root cause here is that all strings returned by
out JSON RPC handler are HTML escaped for security reasons.

In this particular case we're trying to open a new tab/window in
the browser with a URL to an external bug tracker. This URL very
likely contains query parameters and & is a valid character in
this case.

The original bug report says JIRA can't deal with &amp;.
Interestingly GitHub can parse the escaped format in the same way
as the unescaped one!