Try harder to restore original settings. Fixes #991
root cause is that settings are cached and after override
MENU_ITEMS is modified. In case of intermediary failures it was
not restored and other pages were showing bogus menu items.

This seems to resolve the problem although it's now very hard to
reproduce this problem reliably, hence no actual test.