Add tests for presense of mysql/psql binaries in container
because we tend to install only the libraries for the DB drivers.

Last time when we switched to MariaDB 10.3 we broke this!

These binaries are the main DB cli tool which is also used by
`manage.py dbshell`, which in turn is referenced in some use-cases
around truncating and restoring data into existing databases!