docs: Remove mention about superuser account creation in webUI
this depends on the email backend being configured, b/c it needs
to verify the email address, which is something that many users
will struggle with or not be considered as a 1st step config task.

We should move toward providing an initial configuration via webUI
experience + a single cli entrypoint to be used by site admins!