Add migrations_order command to help test rollbacks
the output is the same as ./manage.py migrate --plan but more
suitable for machine parsing. Format is <app_name.migration_name>,
no colors.

This command is intended for testing purposes but it can be useful
in case you are rolling back to an older version and need to see
in which order to rollback the migrations.