Handle history when importing objects with ID field set
in some cases, e.g. custom data restore, we're saving objects with
all of their fields set, including the PK/ID field. Without this
patch instance.pk will be present in .pre_save() but the actual
record doesn't exist in the DB yet hence an exception will be raised.

With this change instance.previous will be None in case the object
doesn't exist in the DB and the code in .post_save() will take care
to execute the diff_objects() function only when instance.previous
isn't None.