Move post_save.send() from bugs.views to comments.add_comment()
any API/method using this helper will also trigger post_save()
handlers for the referenced object class which, among other things,
will take care to reassign the attachments to the same object to
which the comments are added.

This allows cleaner organization of attachments which came via the
rich text editor (usually images).