Send unescaped input to Markdown.render() API via JavaScript
when rendering Markdown into HTML unescape the input b/c it
may be coming from another API call and could be preescaped.
This is the case for example when adding comments!

The BE rendering functionality will take care to remove any
characters that are deemed unsafe!

If we don't do this then some inputs already contain escaped
sequences, e.g. quotes in code blocks which don't appear well
when rendered as HTML.