Refactor TestCaseStatus.get_confirmed() method
NOTE: the previous implementation seems to work fine for views
and with API calls when executed live. However it causes one test
to fail b/c it can't find the 'CONFIRMED' status!

This implementation is not the best one but is functionally
equivalent and seems to work. Also all of this will be refactored
as part of #1932 very soon.