Close dropdown when bulk-actions are activated and don't follow #
NOTE: for 2 existing menu items the handler function syntax has
been updated from `() => {}` to `function() {}` !!!

In the first case $(this) referred to Window and couldn't find the
parent dropdown menu in order to close it. IDK why this happens,
there shouldn't be any difference between the two syntaxes.

CC @asankov