Add change status buttons next to save comment btn
Expand new row on status changes

change `iElement` click handler to vanilla JS function, so that we have
access to the caller element.

Change `reloadRowFor` to call `treeViewBind` only for the container that
we are reloading the row for. Otherwise we have already expanded the
next row, but `treeViewBind` adds the `hidden` class to the details.
The other options would be to wait for the change status call to succeed
and only then expand the next row, but that is slow and bad UX.

Add comment upon click of status button