Properly display the bug icon when adding links to bugs
use removeClass() instead of toggleClass(). Fixes a bug where
there are already bugs attached to this TE and the icon is shown,
either from initial load or from previous bugs added via the menu
and then the user adds one more hyperlink which caused the icon
to be hidden (because class 'hidden' was toggled, e.g. added back).