tests: Change existing Azure Boards issue to 717
the root cause of failing tests with issue #1 is that Azure API
by default returns 200 comments and the waiting mechanism we have
in place fails b/c initial_count == current_count == 200.

As a result the loop executes several hundred times and Azure
detects that the breaks the connection with us.