Pass DataTables `callback` to `pre_process_data` b/c async
in case pre_process_data() needs to make other PRC calls we want
them to be async which means pre_process_data() needs to instruct
DataTables that information is ready for rendering!

WARNING: this breaks compatibility with previous implementation!

TODO: all places where we call dataTableJsonRPC() and augment the
information before rendering it MUST be updated!