Refactor TestPlan search page rendering
because of how additional information is now fetched we can
fetch information in bulk, e.g. all tags, instead of fetching
them one by one! This reduces the total number of RPC queries
because lots of this information is duplicate. In turn this
speeds up loading times on this page. Refs #1014