Pass OBJECT_MENU_ITEMS as context variable. Fixes #991
the previous way of overriding MENU_ITEMS has proven to be fragile
and despite us trying to handle failures there are still cases in
which multiple object menus appear in the navbar. Usually this
happens after experiencing some sort of crash on the backend.

This commit teaches navbar.html to search for a new context variable
which would represent the object-level menus. Any views that care
to display those can pass OBJECT_MENU_ITEMS in their context in the
same format as MENU_ITEMS!