Document performance for TestCase.filter/TestRun.filter methods. Closes #1173
worst case scenario I get around 13 seconds to generate and transfer
5 MB for 10k records. The original issue says 8 seconds for 5000 records
but from the screenshots it is evident that the data size is 25.4 KB.

This is nowhere near half of what we show in this experiment so my
inclination is that the reporter of #1173 is using a very slow network.
This is also supported by the predominant green bar in the picture which
represents Download time in Firefox's metrics.