Add `eslintConfig` in `package.json`
this config excludes the `no-undef` rules.
Currently, a lot of the errors we get are violating this rule.
This will be fixed when we integrate Webpack in our workflow.
Until then, there is no point in showing these messages, because
they only clutter the output and prevent us from seeing the actual
warnings and errors.