Use TestCase + TestRun properties when generating test executions
in the case where environment(s) is specified for the new TestRun all of
its properties apply to all of the contained test executions.

In the case where you have good variability in your environment and
especially combined with parameterized tests using the `pairwise`
generation algorithm will reduce the testing matrix significantly.

If this isn't desired then use the `full` matrix and decide which
executions need to be tested/re-tested.