Restore the tech-preview feature notices
I think there's more work that will need to be done around how
environment properties are selected for a TestRun and how they are
combined together with TestCase properties, especially when reducing the
testing matrix!

In some use-cases it seems like Environment is the same as group of
properties on a global level, while in others it's more like a group of
tags with only a few of the properties having multiple values, which
kind of defeats the purpose of having global test parameters in the
first place.