Convert values to array when updating Version/Build. Fixes #2704.
On pages where a single selection can be made tha value is a string,
while on pages with multiple selection the value is an array. This
causes the API to not return any values b/c the selector

	product__in="string"

doesn't work on Postgres.

Making sure that we always pass array as an argument to the ORM selector
works on all database backends.