Don't send duplicate emails when editting bugs. Fixes #2782
the root cause here is that the post-save signal is called multiple
times from various triggers. The trigger which we're interested in is
when post-save signal is triggered by the add_comment() function b/c
that's where the important information about bugs is stored.

Move tests for email sending from model to view tests!