Optimize Status Matrix 20x and reduce it down to 7 queries. Refs #1801
by refactoring Telemetry.status_matrix() API method and it's front-end
handling. Instead of aggregating the results in pure Python return most
of the data in separate fields inside the response data structure.

On the FE update which fields are used when rendering in order to match
the new data structure.